chore: move schema under src
